fix: check if setup workdir is set before creating

+++ cmd/skiff-core/command_setup.go
@@ -2,6 +2,7 @@ package main
 
 import (
 	"os"
+	"strings"
 
 	"github.com/paralin/skiff-core/setup"
 	"github.com/urfave/cli"
@@ -38,14 +39,17 @@ var SetupCommands cli.Commands = []cli.Command{
 				return cli.NewExitError("Unable to parse config: "+err.Error(), 1)
 			}
 
-			if _, err := os.Stat(setupArgs.WorkDir); err != nil {
-				if os.IsNotExist(err) {
-					// if we created the dir, remove it afterwards.
-					defer os.RemoveAll(setupArgs.WorkDir)
-				}
-				err = os.Mkdir(setupArgs.WorkDir, 0755)
-				if err != nil {
-					return cli.NewExitError("Unable to create working directory: "+err.Error(), 1)
+			setupArgs.WorkDir = strings.TrimSpace(setupArgs.WorkDir)
+			if setupArgs.WorkDir != "" {
+				if _, err := os.Stat(setupArgs.WorkDir); err != nil {
+					if os.IsNotExist(err) {
+						// if we created the dir, remove it afterwards.
+						defer os.RemoveAll(setupArgs.WorkDir)
+					}
+					err = os.Mkdir(setupArgs.WorkDir, 0755)
+					if err != nil {
+						return cli.NewExitError("Unable to create working directory: "+err.Error(), 1)
+					}
 				}
 			}
